% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeSiteEff_l2_par.R
\name{CodeSiteEff_l2_par}
\alias{CodeSiteEff_l2_par}
\title{Function Used To Estimate Code-Site Effects Parallelly}
\usage{
CodeSiteEff_l2_par(
  S.1,
  S.2,
  n1,
  n2,
  U.1,
  U.2,
  V.1,
  V.2,
  delta.int,
  lambda.delta,
  p,
  common_codes,
  n.common,
  n.core
)
}
\arguments{
\item{S.1}{SPPMI from the source site}

\item{S.2}{SPPMI from the target site}

\item{n1}{the number of codes from the source site}

\item{n2}{the number of codes from the target site}

\item{U.1}{the left embeddings (left singular vectors times the square root of the singular values) from the source site}

\item{U.2}{the left embeddings (left singular vectors times the square root of the singular values) from the target site}

\item{V.1}{the right embeddings (right singular vectors times the square root of the singular values) from the source site}

\item{V.2}{the right embeddings (right singular vectors times the square root of the singular values) from the target site}

\item{delta.int}{the initial estimator for the code-site effect}

\item{lambda.delta}{the tuning parameter controls the intensity of penalization on the code-site effects}

\item{p}{the length of an embedding}

\item{common_codes}{the list of overlapping codes}

\item{n.common}{the number of overlapping codes}

\item{n.core}{the number of cored used for parallel computation}
}
\value{
The output for the estimation of code-site effects
}
\description{
Function Used To Estimate Code-Site Effects Parallelly
}
