% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{download_dataset}
\alias{download_dataset}
\title{Download Datasets from Zenodo}
\usage{
download_dataset(dataset, destfile = paste0(dataset, ".rda"))
}
\arguments{
\item{dataset}{The name of the dataset to download (e.g., "pairs.rel.CV").}

\item{destfile}{The destination file path where the dataset will be saved. Defaults to the current working directory.}
}
\value{
A message confirming the dataset has been downloaded.
}
\description{
This function downloads large datasets hosted on Zenodo for use with the MUGS package.
}
\examples{
\dontrun{
  download_dataset("pairs.rel.CV", "data/pairs.rel.CV.rda")
}
}
