% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{download_example_data}
\alias{download_example_data}
\title{Download and Load Example Data from Zenodo}
\usage{
download_example_data(
  file,
  cache_dir = tools::R_user_dir("MUGS", which = "cache")
)
}
\arguments{
\item{file}{Name of the .rda/.Rdata file to download.}

\item{cache_dir}{Directory to store the downloaded data locally.}
}
\value{
Loads the dataset into the global environment.
}
\description{
Download and Load Example Data from Zenodo
}
