% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.sim.R
\name{evaluation.sim}
\alias{evaluation.sim}
\title{Function Used For Tuning And Evaluation}
\usage{
evaluation.sim(pairs.rel, U)
}
\arguments{
\item{pairs.rel}{the known code-code pairs}

\item{U}{the code embedding matrix}
}
\value{
The output of tuning and evaluation
}
\description{
Function Used For Tuning And Evaluation
}
\examples{
pairs_rel_EV_data <- system.file("extdata", "pairs.rel.EV.Rdata", package = "MUGS")
load(pairs_rel_EV_data)
U2_data <- system.file("extdata", "U.2.Rdata", package = "MUGS")
load(U2_data)
evaluation.sim(pairs.rel.EV, U.2)
}
