% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embed.R
\name{get_embed}
\alias{get_embed}
\title{Function For Getting Embedding From SVD}
\usage{
get_embed(mysvd, d = 2000, normalize = TRUE)
}
\arguments{
\item{mysvd}{the (managed) svd result (adding an element with 'names')}

\item{d}{dim of the final embedding}

\item{normalize}{if the output embeddings have l2 norm equal to 1}
}
\value{
The embedding from SVD
}
\description{
Function For Getting Embedding From SVD
}
