% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{load_pairs_rel_CV}
\alias{load_pairs_rel_CV}
\title{Load pairs.rel.CV Dataset}
\usage{
load_pairs_rel_CV(filepath = NULL)
}
\arguments{
\item{filepath}{(optional) Path to the downloaded `.Rdata` file.
If `NULL`, the file will be downloaded automatically.}
}
\value{
The loaded dataset.
}
\description{
This function loads the `pairs.rel.CV` dataset into R.
}
\examples{
\dontrun{
pairs.rel.CV <- load_pairs_rel_CV()
}
}
