% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{load_S_1}
\alias{load_S_1}
\title{Load S.1 Dataset}
\usage{
load_S_1(filepath = NULL)
}
\arguments{
\item{filepath}{(optional) Path to the downloaded `.Rdata` file.
If `NULL`, the file will be downloaded automatically.}
}
\value{
The loaded dataset.
}
\description{
This function loads the `S.1` dataset into R.
}
\examples{
\dontrun{
S.1 <- load_S_1()
}
}
