% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{load_S_2}
\alias{load_S_2}
\title{Load S.2 Dataset}
\usage{
load_S_2(filepath = NULL)
}
\arguments{
\item{filepath}{(optional) Path to the downloaded `.Rdata` file.
If `NULL`, the file will be downloaded automatically.}
}
\value{
The loaded dataset.
}
\description{
This function loads the `S.2` dataset into R.
}
\examples{
\dontrun{
S.2 <- load_S_2()
}
}
