% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{load_U_1}
\alias{load_U_1}
\title{Load U.1 Dataset}
\usage{
load_U_1(filepath = NULL)
}
\arguments{
\item{filepath}{(optional) Path to the downloaded `.Rdata` file.
If `NULL`, the file will be downloaded automatically.}
}
\value{
The loaded dataset.
}
\description{
This function loads the `U.1` dataset into R.
}
\examples{
\dontrun{
U.1 <- load_U_1()
}
}
