% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{load_X_group_target}
\alias{load_X_group_target}
\title{Load X.group.target Dataset}
\usage{
load_X_group_target(filepath = NULL)
}
\arguments{
\item{filepath}{(optional) Path to the downloaded `.Rdata` file.
If `NULL`, the file will be downloaded automatically.}
}
\value{
The loaded dataset.
}
\description{
This function loads the `X.group.target` dataset into R.
}
\examples{
\dontrun{
X.group.target <- load_X_group_target()
}
}
