% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUGS.R
\name{MUGS}
\alias{MUGS}
\title{Main function for MUGS algorithm}
\usage{
MUGS(
  TUNE = F,
  Eva = T,
  Lambda = c(10),
  Lambda.delta = c(1000),
  n.core = 4,
  tol = 1,
  seed = 1,
  S.1 = NULL,
  S.2 = NULL,
  X.group.source = NULL,
  X.group.target = NULL,
  pairs.rel.CV = NULL,
  pairs.rel.EV = NULL,
  p = 100,
  n.group = 400
)
}
\arguments{
\item{TUNE}{Logical value indicating whether the function should tune parameters 'TRUE' or use predefined parameters 'FALSE'.}

\item{Eva}{Logical value indicating whether to perform evaluation (TRUE) or skip it (FALSE).}

\item{Lambda}{The candidate values for the tuning parameter controls the intensity of penalization on the code effects.}

\item{Lambda.delta}{The candidate values for the tuning parameter controls the intensity of penalization on the code-site effects.}

\item{n.core}{Integer specifying the number of cores to use for parallel processing.}

\item{tol}{Numeric value representing the tolerance level for convergence in the algorithm.}

\item{seed}{Integer used to set the seed for random number generation, ensuring reproducibility of the simulated data or any stochastic process within the algorithm.}

\item{S.1}{The SPPMI matrix from site 1.}

\item{S.2}{The SPPMI matrix from site 2.}

\item{X.group.source}{The dummy matrix on the group structure of codes at site 1.}

\item{X.group.target}{The dummy matrix on the group structure of codes at site 2.}

\item{pairs.rel.CV}{Code-code pairs used for tuning via cross validation}

\item{pairs.rel.EV}{Code-code pairs used for evaluation}

\item{p}{Integer indicating the length of embeddings.}

\item{n.group}{The number of groups.}
}
\value{
The final result
}
\description{
Main function for MUGS algorithm
}
\examples{
S1_data <- system.file("extdata", "S.1.Rdata", package = "MUGS")
load(S1_data)
S2_data <- system.file("extdata", "S.2.Rdata", package = "MUGS")
load(S2_data)
pairs_rel_CV_data <- system.file("extdata", "pairs.rel.CV.Rdata", package = "MUGS")
load(pairs_rel_CV_data)
pairs_rel_EV_data <- system.file("extdata", "pairs.rel.EV.Rdata", package = "MUGS")
load(pairs_rel_EV_data)
X_group_source_data <- system.file("extdata", "X.group.source.Rdata", package = "MUGS")
load(X_group_source_data)
X_group_target_data <- system.file("extdata", "X.group.target.Rdata", package = "MUGS")
load(X_group_target_data)

}
