% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{pairs.rel.EV}
\alias{pairs.rel.EV}
\title{pairs.rel.EV Dataset}
\format{
A data frame with multiple columns:
\describe{
  \item{col}{Integer representing the column index of a pair.}
  \item{row}{Integer representing the row index of a pair.}
  \item{type}{Character string indicating the type of data (e.g., "validation").}
}
}
\usage{
pairs.rel.EV
}
\description{
A data frame containing evaluation pairs for relative comparisons.
}
\keyword{datasets}
