% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{S.1}
\alias{S.1}
\title{S.1 Dataset}
\format{
A matrix with 2000 rows and 10 columns:
\describe{
  \item{Row Names}{Unique identifiers for each row.}
  \item{Columns}{Numeric values representing SPPMI data.}
}
}
\usage{
S.1
}
\description{
A matrix containing SPPMI data from the source site.
This dataset is used as input for analysis in the package.
}
\keyword{datasets}
