% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{U.1}
\alias{U.1}
\title{U.1 Dataset}
\format{
A matrix with 2000 rows and 10 columns:
\describe{
  \item{Row Names}{Unique identifiers for each row.}
  \item{Columns}{Numeric values representing embeddings.}
}
}
\usage{
U.1
}
\description{
A matrix containing left embeddings from the source site.
These embeddings are used for embedding-based computations.
}
\keyword{datasets}
