% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{X.group.source}
\alias{X.group.source}
\title{X.group.source Dataset}
\format{
A matrix with 2000 rows and 50 columns:
\describe{
  \item{Rows}{Entities at the source site.}
  \item{Columns}{Binary values (0 or 1) indicating group membership.}
}
}
\usage{
X.group.source
}
\description{
A matrix containing group structures at the source site.
It represents binary group membership of entities at the source.
}
\keyword{datasets}
