% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{X.group.target}
\alias{X.group.target}
\title{X.group.target Dataset}
\format{
A matrix with 2000 rows and 50 columns:
\describe{
  \item{Rows}{Entities at the target site.}
  \item{Columns}{Binary values (0 or 1) indicating group membership.}
}
}
\usage{
X.group.target
}
\description{
A matrix containing group structures at the target site.
It represents binary group membership of entities at the target.
}
\keyword{datasets}
