% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Datasets.R
\name{download_example_data}
\alias{download_example_data}
\title{Download and Load Example Data from Zenodo}
\usage{
download_example_data(file, destdir = tempdir())
}
\arguments{
\item{file}{Name of the .Rdata file to download (e.g., "S.1.Rdata").}

\item{destdir}{Directory to store the downloaded data. Defaults to a temporary directory.}
}
\value{
A list containing the loaded dataset.
}
\description{
Download and Load Example Data from Zenodo
}
\examples{
\donttest{
data_list <- download_example_data("S.1.Rdata")
S.1 <- data_list[["S.1"]]
}
}
