% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1.R
\docType{data}
\name{S.1}
\alias{S.1}
\title{S.1}
\format{
A matrix with 2000 rows and 50 columns.
}
\source{
Generated as part of the MUGS package for demonstrating usage.
}
\usage{
S.1
}
\description{
A sample dataset used in the MUGS package.
}
\examples{
data(S.1)
}
\keyword{datasets}
